requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.0
numpy==1.24.3
torch==2.0.1
transformers==4.30.0
pymysql==1.1.0
